add query.py
